moved existing shangrla functions to core
reconfigured package with only pyproject.toml
