+++ pyproject.toml
@@ -4,3 +4,51 @@ requires = [
     "wheel"
 ]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+exclude = [
+  "Examples"
+]
+
+[project]
+name = "shangrla"
+version = "0.0.1"
+authors = [
+  {name = "Philip B. Stark", email = "pbstark@berkeley.edu"},
+  {name = "Vanessa Teague"},
+  {name = "Michelle Blom"}
+]
+maintainers = [
+  {name = "Philip B. Stark", email = "pbstark@berkeley.edu"},
+]
+description = "A Python library for risk-limiting audits"
+readme = "README.md"
+license = {file = "LICENSE.md"}
+keywords = ["risk-limiting audits", "nonparametric tests"]
+classifiers = [
+  "Programming Language :: Python :: 3.10",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent"
+]
+
+dependencies = [
+    "numpy",
+    "scipy >= 0.17.1",
+    "cryptorandom >= 0.2",
+    "svgling",
+    "colorama"
+]
+
+[project.optional-dependencies]
+examples = [
+    "matplotlib",
+    "pandas"
+]
+test = [
+    "pandas",
+    "pytest",
+    "pytest-cov",
+    "coverage",
+    "flake8",
+    "openpyxl"
+]
