--- setup.cfg
@@ -1,54 +1,0 @@
-# Declarative configuration for setup.py
-
-# For more details on this format, see the official docs here:
-# https://setuptools.pypa.io/en/latest/userguide/declarative_config.html
-# And a useful annotated template can be found here:
-# https://gist.github.com/althonos/6914b896789d3f2078d1e6237642c35c
-
-[metadata]
-name             = shangrla
-version          = attr: shangrla.__version__
-author           = Philip B. Stark, Vanessa Teague, Michelle Blom
-author_email     = pbstark@berkeley.edu
-description      = A Python library for risk-limiting audits
-long_description = file: README.md, LICENSE.md
-long_description_content_type = text/markdown
-keywords         = risk-limiting audits, nonparametric tests
-license          = MIT 
-classifiers      =
-	Programming Language :: Python :: 3
-    License :: OSI Approved :: MIT License
-    Operating System :: OS Independent
-
-[options]
-include_package_data = True
-packages = find:
-python_requires  = >= 3.10.4
-install_requires =
-    tqdm
-    numpy
-    scipy >= 0.17.1
-    pandas
-    matplotlib
-    cryptorandom >= 0.2
-    svgling
-    svgwrite
-    colorama
-    xlrd
-    openpyxl
-    ipywidgets
-
-[options.packages.find]
-exclude =
-    examples*
-    docs*
-
-
-[options.extras_require]  # in setup.cfg
-test =
-    pytest
-    pytest-cov
-    coverage
-    flake8
-examples = 
-    pandas
