+++ tests/Core/test_Dominion.py
@@ -16,10 +16,10 @@ from cryptorandom.sample import random_permutation
 from cryptorandom.sample import sample_by_index
 
 
-from shangrla.Audit import Audit, Assertion, Assorter, Contest, CVR, Stratum
-from shangrla.NonnegMean import NonnegMean
-from shangrla.Dominion import Dominion
-from shangrla.Hart import Hart
+from shangrla.core.Audit import Audit, Assertion, Assorter, Contest, CVR, Stratum
+from shangrla.core.NonnegMean import NonnegMean
+from shangrla.core.Dominion import Dominion
+from shangrla.core.Hart import Hart
 
 ##########################################################################################
 
@@ -29,7 +29,7 @@ class TestDominion:
         Test reading the older Dominion CVR format
         """
         cvr_list = Dominion.read_cvrs(
-            "tests/Data/Dominion_CVRs/test_5.2.18.2.Dominion.json", pool_groups = [1]
+            "tests/Core/Data/Dominion_CVRs/test_5.2.18.2.Dominion.json", pool_groups = [1]
         )
         assert len(cvr_list) == 2
         cvr_1, cvr_2 = cvr_list
@@ -59,7 +59,7 @@ class TestDominion:
         Ingests the older Dominion CVR format
         """
         cvr_list = Dominion.read_cvrs(
-            "tests/Data/Dominion_CVRs/test_5.2.18.2.Dominion.json",
+            "tests/Core/Data/Dominion_CVRs/test_5.2.18.2.Dominion.json",
             use_adjudicated=True,
         )
         assert len(cvr_list) == 2
@@ -89,7 +89,7 @@ class TestDominion:
         Test reading the newer Dominion CVR format
         """
         cvr_list = Dominion.read_cvrs(
-            "tests/Data/Dominion_CVRs/test_5.10.50.85.Dominion.json"
+            "tests/Core/Data/Dominion_CVRs/test_5.10.50.85.Dominion.json"
         )
         assert len(cvr_list) == 2
         cvr_1, cvr_2 = cvr_list
@@ -112,7 +112,7 @@ class TestDominion:
         Ingests the newer Dominion CVR format
         """
         cvr_list = Dominion.read_cvrs(
-            "tests/Data/Dominion_CVRs/test_5.10.50.85.Dominion.json",
+            "tests/Core/Data/Dominion_CVRs/test_5.10.50.85.Dominion.json",
             use_adjudicated=True
         )
         assert len(cvr_list) == 2
@@ -137,7 +137,7 @@ class TestDominion:
         Test reading the newer Dominion CVR format, selecting VBM results only
         """
         cvr_list = Dominion.read_cvrs(
-            "tests/Data/Dominion_CVRs/test_5.10.50.85.Dominion.json",
+            "tests/Core/Data/Dominion_CVRs/test_5.10.50.85.Dominion.json",
             include_groups=[2]
         )
         assert len(cvr_list) == 1
@@ -154,7 +154,7 @@ class TestDominion:
         Tests the convenience function for reading all CVRs from a given directory
         """
         cvr_list = Dominion.read_cvrs_directory(
-            "tests/Data/Dominion_CVRs/CVR_Export", pool_groups=[2]
+            "tests/Core/Data/Dominion_CVRs/CVR_Export", pool_groups=[2]
         )
         assert len(cvr_list) == 2
         cvr_1, cvr_2 = cvr_list
